+++ include/D3D12Helper/D3D12Interop/D3D12Interop.hpp
@@ -1,4 +1,7 @@
 #pragma once
 #include <D3D12Helper/D3D12Core/D3D12Fence.hpp>
+#include <D3D12Helper/D3D12Core/D3D12SharedResource.hpp>
 #include <D3D12Helper/D3D12Framework/D3D12Resource.hpp>
 #include <D3D12Helper/D3D12Framework/D3D12Helpers.hpp>
+#include <D3D12Helper/D3D12Interop/D3D12SharedHandle.hpp>
+#include <D3D12Helper/D3D12Interop/D3D12SharedResource.hpp>